move unit test to TranslatorSyntheticTest
